Add ts stub for useTaskStore
